refactor: retarget agent user snapshot install owner
